# Optimize `Time.at_with_coercion`